intra2/requirements.txt
amqp==1.4.7
anyjson==0.3.3
billiard==3.3.0.21
celery==3.1.19
dj-database-url==0.3.0
Django==1.9
django-celery==3.1.17
django-cors-headers==1.0.0
django-enumerify==0.3.3
django-extensions==1.5.9
django-filter==0.11.0
django-redis-cache==1.6.4
django-redis-sessions==0.5.0
django-rest-swagger==0.3.4
djangorestframework==3.3.1
djangorestframework-jwt==1.7.2
djrill==1.3.0
isoweek==1.3.0
kombu==3.0.29
pprintpp==0.2.1
psycopg2==2.6
PyJWT==1.4.0
pytz==2015.7
PyYAML==3.11
raven==5.3.1
redis==2.10.5
requests==2.8.1
six==1.10.0
tabulate==0.7.5
Werkzeug==0.9.6
xlrd==0.9.3